add hwaccel support, tests not finished

+++ include/av/Stream.hpp
@@ -26,14 +26,16 @@ public:
 	/**
 	 * @return A pointer to the contained `AVStream`.
 	 * @warning **Do not free/delete the returned pointer.**
-	 * It belongs to and is managed by an `AVFormatContext`, which is wrapped by `MediaReader`.
+	 * It belongs to and is managed by an `AVFormatContext`, which is wrapped by
+	 * `MediaReader`.
 	 */
 	AVStream *get() { return _s; }
 
 	/**
 	 * @return A read-only pointer to the contained `AVStream`.
 	 * @warning **Do not free/delete the returned pointer.**
-	 * It belongs to and is managed by an `AVFormatContext`, which is wrapped by `MediaReader`.
+	 * It belongs to and is managed by an `AVFormatContext`, which is wrapped by
+	 * `MediaReader`.
 	 */
 	const AVStream *get() const { return _s; }
 
@@ -41,7 +43,8 @@ public:
 	 * Access fields the contained `AVStream`.
 	 * @return A pointer to the contained `AVStream`.
 	 * @warning **Do not free/delete the returned pointer.**
-	 * It belongs to and is managed by an `AVFormatContext`, which is wrapped by `MediaReader`.
+	 * It belongs to and is managed by an `AVFormatContext`, which is wrapped by
+	 * `MediaReader`.
 	 */
 	AVStream *operator->() { return _s; }
 
@@ -49,15 +52,21 @@ public:
 	 * Access fields the contained `AVStream`.
 	 * @return A read-only pointer to the contained `AVStream`.
 	 * @warning **Do not free/delete the returned pointer.**
-	 * It belongs to and is managed by an `AVFormatContext`, which is wrapped by `MediaReader`.
+	 * It belongs to and is managed by an `AVFormatContext`, which is wrapped by
+	 * `MediaReader`.
 	 */
 	const AVStream *operator->() const { return _s; }
 
 	/**
-	 * Wrapper over `av_dict_get` using the contained `AVStream`'s `metadata` field.
-	 * @returns The value associated with `key` in stream `i`, or `NULL` if no entry exists for `key`.
+	 * Wrapper over `av_dict_get` using the contained `AVStream`'s `metadata`
+	 * field.
+	 * @returns The value associated with `key` in stream `i`, or `NULL` if no
+	 * entry exists for `key`.
 	 */
-	const char *metadata(const char *const key, const AVDictionaryEntry *prev = NULL, const int flags = 0) const
+	const char *metadata(
+		const char *const key,
+		const AVDictionaryEntry *prev = NULL,
+		const int flags = 0) const
 	{
 		const auto entry = av_dict_get(_s->metadata, key, prev, flags);
 		return entry ? entry->value : NULL;
@@ -66,12 +75,18 @@ public:
 	/**
 	 * @return The duration of the stream, in seconds.
 	 */
-	double duration_sec() const { return _s->duration * av_q2d(_s->time_base); }
+	double duration_sec() const
+	{
+		return static_cast<double>(_s->duration) * av_q2d(_s->time_base);
+	}
 
 	/**
 	 * @return The total number of audio samples (per channel) in this stream.
 	 */
-	int samples() const { return duration_sec() * sample_rate(); }
+	int samples() const
+	{
+		return static_cast<int>(duration_sec() * sample_rate());
+	}
 
 	/**
 	 * @return The number of audio channels in this stream.
@@ -87,13 +102,20 @@ public:
 	 * @return A `Decoder` using this stream's codec.
 	 * @note Codec parameters are not copied to the context.
 	 */
-	Decoder create_decoder() const { return avcodec_find_decoder(_s->codecpar->codec_id); }
+	Decoder create_decoder() const
+	{
+		return avcodec_find_decoder(_s->codecpar->codec_id);
+	}
 
 	/**
-	 * @return An `Encoder` using this stream's codec identified by `_s->codecpar->codec_id`.
+	 * @return An `Encoder` using this stream's codec identified by
+	 * `_s->codecpar->codec_id`.
 	 * @note Codec parameters are not copied to the context.
 	 */
-	Encoder create_encoder() const { return avcodec_find_encoder(_s->codecpar->codec_id); }
+	Encoder create_encoder() const
+	{
+		return avcodec_find_encoder(_s->codecpar->codec_id);
+	}
 };
 
 } // namespace av
